Document system to bypass Twitter rate limits
This post details the author's experience in creating a system to bypass Twitter's rate limits, explaining the technical architecture and strategies used, as well as the eventual consequences of those actions. It also outlines the lessons learned and the transition to a more sustainable approach in the second version of the system.